go_vendor_license: work around bug with old tomlkit versions
Older versions of Fedora still have this bug, and the new way makes much
more sense when we're only setting a single key in the dictionary.
This was originally written like this when the plan was to add
additional items to the `write_config_data` dictionary.

Relates: https://github.com/python-poetry/tomlkit/issues/331