license: allow --autofill when --prompt is not passed
We should still call `fill_missing_licenses()` even if `prompt` is disabled.
The rest of the code was changed to allow either or both options to be
passed (or neither to be passed — in which case, `fill_missing_licenses`
isn't called at all, as this block would still evaluate to False), but
this condition was still broken.