Make zstarfile dependency optional and use tar filter
This vendors the extra helper functions we use from zstarfile and makes
optional the usage of the ZSTarfile class that adds support for lz4 and
zstandard (on Python < 3.14 where it's not already supported).

This also modifies the code to use the tar filter by default when
unpacking archives.

Fixes: https://gitlab.com/fedora/sigs/go/go-vendor-tools/-/issues/84
Relates: https://gitlab.com/fedora/sigs/go/go-vendor-tools/-/issues/3